services/backend/requirements.txt
# Main Application Requirements
fastapi==0.104.1
celery==5.3.4
flower==2.0.1  # Celery monitoring
redis==5.0.1
motor==3.3.2
pymongo==4.6.1
beanie==1.23.6
langchain==0.3.0
langchain-openai==0.3.0
pydantic==2.8.0
httpx==0.26.0
boto3==1.34.34
structlog==24.1.0
python-multipart==0.0.6
websockets==12.0
uvicorn==0.27.0
python-dotenv==1.0.1
psutil==5.9.8
ffmpeg-python==0.2.0
